AttentionAgent/requirements.txt
Cython==0.29.14
gin-config==0.1.4
protobuf==3.7.1
grpcio==1.20.1
grpcio-tools==1.20.1
gym[box2d]==0.15.3
pyglet==1.3.2
numpy==1.16.3
pandas==0.24.2
seaborn==0.9.0
matplotlib==3.0.0
cma==2.7.0
torch==1.3.0
torchvision==0.4.1
pillow==6.2.1
google-api-core==1.14.3
google-auth==1.7.1
google-cloud-core==1.0.3
google-cloud-storage==1.23.0
google-pasta==0.1.7
google-resumable-media==0.5.0
googleapis-common-protos==1.6.0
opencv-python==4.1.2.30
#ppaquette_gym_doom
